Added static elements and tables
Added a placeholder pdf for the 'paperView' webpage (still isn't working, however I just need to figure out how to embed it properly)... gradeView/myBinder has a temporary static table so the page doesn't look as barren and is presentable. Still need to fix registration form and allow for communication between the site and our db.